Add a configuration flag to completely disable printing backtraces
In general, backtraces are really long and noisy, and using logs as a
means of diagnosing bugs instead of some sort of exception tracker is
never a particularly good idea.

We could move away from passing backtraces to chalk-log entirely, but
the automatic rendering of error_class and error are useful.