Add global default log level configuration to Chalk::Log
Currently, Chalk::Log instantiates a logger for each module when
`log` is first called on that module. At instantiation, it sets the
log level to `configatron.chalk.log.default_level`. This means that
if you want to set a different global log level, you either need
to do it before `log` is called or do it for each module individually.
Both are error prone and difficult.

This adds a new logger at the root of the logging hierarchy where we
set the configatron default log level once. You can override the
global default at any time with `Chalk::Log.level = <level>`. All modules
that haven't set an explicit log level (or have explicitly set their
level to `nil`) will respect the global default.

This also fixes a bug whereby `log` in anonymous classes and modules
was mutating the log level on the root logger.