Respect precedence of contextual info
Nested contextual info should take precedence over an outer block when
they are using the same key. For example, in the following block, the
contextual info should be "inner" when the logger is called.

```
log.with_contextual_info(shared_key: "outer") do
  log.with_contextual_info(shared_key: "inner") do
    log.info("message")
  end
end
```

Incorrect ordering on a hash merge was causing this order not to be
respected, resulting in the topmost key cascading through and
overwriting narrower scopes. This commit reverses that merge order to
respect the set precedence.